projetos-nao-iniciados/newsletter-tecnologia-agno/requirements.txt
agno==2.1.0
openai>=1.107.1
python-dotenv>=1.1.1
tavily-python>=0.7.12